+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Improved selectors and sorting
+- In xls doc, verifying reviews are returned in sorted order
+
 ### Added
 
 - For verifiedPurchaser testcase, added HasShopperReviewed intercept